feat: Adding new EC2 instances and its respective configuration (#1)
* feat: Adding new EC2 instances and its respective configuration

In this commit, various EC2 instances were added in accordance with the pipeline project
The respective instances are being configured with their necessary ports and access from
particular security groups only
The bastion host was also added and it is the only way to access internal servers in the same subnet
since there is no public IP assigned to them

* fix: removing unecessary commented code

* fix: terraform fmt -write -recursive command was applyed to fix all deprecated formatting

* fix: Implementing code DRY

implemented the use of code blocks to avoid code repetition. It was done using local and global variables. To combine both, "merge" was used.

* Update 0001-pipeline/terraform/variables.tf

Co-authored-by: Enrico Alvarenga <enricomalvarenga@gmail.com>

* Update 0001-pipeline/terraform/network_and_security.tf

Co-authored-by: Enrico Alvarenga <enricomalvarenga@gmail.com>

* refact: Implemented various improvement to the code

The code was improved to use code blocks variable and the block type which is part of the DRY implementation requested for this code
changes were made to the tags and the security groups
deprecated/invalid formats were also fixed in this refactor.

* fix: removing the use of dynamic blocks for security group configuration

we decided to not use dynamic blocks for security groups configuration
since security groups has its own variation depending on what is needed to be accessed

---------

Co-authored-by: Enrico Alvarenga <enricomalvarenga@gmail.com>