refactor: Created a new workspace to implement the S3 backend
The S3 backend was implemented. The folder backendS3 was created thus a new workspace was created to deploy the S3 bucket.
Both S3 workspace and the pipeline workspace now has their terraform.state saved in the bucket.
The bucket has force_destroy = false and prevent destroy = true to prevent accidental deletion of this bucket.