refactor: Adding security configuration to the EC2 instances
Added all the security groups for the respective EC2 instances
There are 2 subnets, public and private. The public will only have the bastion host
in it and the private will have the remaining instances.
Only the public subnet has the routing table associated to it and thus the internet gateway associated to that routing table.