enhancement: Created new network config and security rules
Createa a new NAT gateway for all the instances in the private subnet to have access to the internet
A Elastic IP is associated with the new NAT gateway. The EIP creation is a precedent for the NAT gateway to be created.
It required a new route table to be created and the private subnet associated to it creating the route through the NAT gateway to the internet