refactor: Adding security configuration to the EC2 instances (#10)
* refactor: Adding security configuration to the EC2 instances

Added all the security groups for the respective EC2 instances
There are 2 subnets, public and private. The public will only have the bastion host
in it and the private will have the remaining instances.
Only the public subnet has the routing table associated to it and thus the internet gateway associated to that routing table.

* fix: fixed the duplicated name on the jfrog-sg

* enhancement: Created new network config and security rules

Createa a new NAT gateway for all the instances in the private subnet to have access to the internet
A Elastic IP is associated with the new NAT gateway. The EIP creation is a precedent for the NAT gateway to be created.
It required a new route table to be created and the private subnet associated to it creating the route through the NAT gateway to the internet